FEAT: Enabled consuming projects to extend DeviceDetectionHashEngine. #397
