+++ Tools/GenerateConfig/GenerateConfig.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="FiftyOne.Pipeline.Web" Version="4.4.126" />
+    <PackageReference Include="FiftyOne.Pipeline.Web" Version="4.5.0-alpha.10" />
     <PackageReference Include="System.Text.Json" Version="6.0.11" />
   </ItemGroup>
