+++ performance-tests/performance-tests.csproj
@@ -10,7 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="FiftyOne.Pipeline.Web" Version="4.4.126" />
+    <PackageReference Include="FiftyOne.Pipeline.Web" Version="4.5.0-alpha.10" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="6.0.36" />
   </ItemGroup>
